Redone links, nav, new style sheet and images (see desc.)
changed the links at bottom of index so they properly scale with window size and don't need tons of extra code and media queries, changed the nav bar to be more aesthetically pleasing, added new font, created new style sheet to work on early life. Most of the work this week has been fixing the styling, so now it's time to focus on getting the bulk of the information onto the pages, Still have a lot to do for the last 2 pages, about 50% done overall.